requirements.txt
Flask==3.1.1
requests==2.32.2
paho-mqtt==2.1.0
